+++ demos/nextjs/app/utils/apiPost.ts
@@ -4,7 +4,7 @@ import axios from 'axios';
 const apiPost = async (endpoint: string, params: any) => {
   const result = await axios.post(`/api${endpoint}`, params, {
     headers: {
-      'content-type': 'application/json',
+      'Content-Type': 'application/json',
     },
   });
   return result.data;