+++ demos/parse-server/public/script.js
@@ -10,7 +10,7 @@ const elBtnPhantom = document.getElementById('auth-phantom');
 const handleApiPost = async (endpoint, params) => {
   const result = await axios.post(`${AUTH_API_URL}/${endpoint}`, params, {
     headers: {
-      'content-type': 'application/json',
+      'Content-Type': 'application/json',
     },
   });