+++ packages/streams/src/mapping/storage/Update.ts
@@ -1,0 +1,9 @@
+export interface Update<Doc = Document> {
+  collectionName: string;
+  document: Doc;
+}
+
+export interface Document {
+  id: string;
+  confirmed: boolean;
+}